#10 #4 Made game run, and player can now move
### CommandMovePlayer
Changed CommandMovePlayer so that it now takes game as an argument. This is to avoid increasing the arguments which KeyboardController needs

### KeyboardController
Changed KeyboardController to now have two dictionaries of keys and commands. KeyPressMappings handles one frame inputs such as using a item or the such. KeyDownMappings handle commands which should function as long as the key is down such as movement.

Additionally added moveCommands to InitializeCommands so that the player can move

### Game1
Changed player from IPlayer to Player along with making it a public {get} property. This is because IPlayer is currently overcomplicating code. If it's needed it can be switched back. It was made public so that KeyboardController can access it with just a game reference

Deleted Kris Sprite since it was clutter

Move Keyboard Controller Initialization to after player was made because movement requires player to be initialized

### Player
Changed player to no longer implement IPlayer because IPlayer just overcomplicates things

Changed PlayerState to PStateMachine along with changing its class to explicitly PlayerStateMachine since that's what it is.

Added ContentManager and SpriteBatch to Player constructor. This should be removed in the future, temp fix so that player actually displays.

### OverworldItemSpriteFactory
Made LoadAllTextures static because that makes more sense. Lupine has shown me the error of my ways

###StateMachineDemo
Mini project to offer an alternative solution to how we should generally handle stateMachines. Incomplete currently, might be able to get it down by meeting start.