#11 Made it so that player can use an item
### Commands
Changed namespaces for consistency

### CommandUseItem
Added UseItemCommand which allows the Player State Machine to use equipment. Currently rough

### KeyboardController
Added a new control for using an item. Assigned it to key D1

### Altered Bow to make use of new spawnedProjectiles system

### IEquipment
Added Enum to IEquipment to keep track of projectiles

### PlayerArrow
Changed it to make use of new spawnedProjectiles system. Additionally changed spawned sprite to potion because arrow sprite wasn't working

### Player
Changed PlayerArrow to instead be a dictionary containing Projectile enums and Projectiles. This way all the spawnedProjectiles can be stored in one dictionary which is then iterated over in draw.

### PlayerStateMachine
Set the active Equipment to the first entry of Player.Equipment for testing purposes

### Util
Added a new method called CardinalToUnitVector so that cardinal to vector translations were safer.

### Projectile
Changed Projectile to use new CardinalToUnitVector method