#11 Added sword swing and fixed boomerang
### CommandSwitchWeaponSprite
Removed due to not being needed

###CommandUseItem
Changed to make use of the EquipmentDictionary so that it didn't rely on index position

### KeyboardController
Reorganized InitializeCommands to have multiple helper functions to better organize the code

Added commands and keybinds to using the sword

### IEquipment
Added enums for weapons for use in the Equipment Dictionary. The current enums are bow,boomerangSack, and swordSheathe

### SwordSheathe
Added a new IEquipment in the vein others for sword spawning

### PlayerBoomerang
Changed playerBoomerang to account for new use of the Move Function

### PlayerSword
Made a new projectile in similar vein to others. It has no movement speed to make it feel swordy

### Projectile
Removed virtual from Update since it shouldn't be being changed

### WaveyArrow
Changed sprite used back to an arrow

### Player
Changed Equipment list to a dictionary consisting of a Weapon Enum and a IEquipment. This allows for easier selection of specific weapons.

Made a function named IntializeEquipment for equipment Initialization

Added SwordSheathe to default list

Removed Functions which had 0 references. Since a lot of them handled with stateMachine stuff which should just be handled in stateMachine

### PlayerStateMachine
Removed activeEquipment because lacked enough purpose along with linkVelocity.

Removed the SwordAttackState since it's been subsumed into ItemUse

Changed UseEquipment to accept a parameter of a Weapon Enum so that it can just use the dictionary rather then needing to have a middle man of activeEquipment. Likely will change in the future tho

### ItemSpriteFactory
Changed source rectangle of Arrow to be the arrow

Changed CreateLinkSprite to create SwordSprite since LinkSprite management happens elsewhere.