Fixed Link's movements and equipment use sprites
###CommandUseItem
Swapped ChangeStateItemUse() and UseEquipment() so that the State Machine had the right state before Link can use an item,

### KeyboardController
Edited HandleStanding to include the item use keys and the if else to be safer.

### Player
Edited Move to just be in charge of changing Link's position, and included a clause to stop link from sliding while in the item frame position. This way player does not handle Link's state and the command calling move does.

### PlayerStateMachine
Edited ChangeDirection to set Link to the standing state

Edited ChangeStateStanding to just do as it says, and commented out the sword attack conditional

Changed GetState to GetCurrentState for more clarity

Added a previous state variable to keep track of Link's state prior to his current state.

Updated Update to keep track of the current state of Link, and prevent any constant changes by checking if his state has changed or not.

### PlayerSpriteFactory
Removed standingLinkSprite, walkingLinkSprite, and itemUseLinkSprite since none would be able to update when called, and in order to update each sprite it would require unnecessary code.

Refactored spriteAtlas to walkingSpriteAtlas for clarity to tell that said sprite atlas is only used during walking frames

Refactored SetXValuesEqual to SetSpritesToFaceSameDirection for clarity

Edited ChangeDirection to set Link's current sprite to be an updated standing sprite facing his current direction

Edited ChangeSpriteStanding, ChangeSpriteWalking, and ChangeSpriteItemUse to update Link's sprite to face his current direction without any issues.