#41 Crappy solution to collisions
### AllCollisionHandler
A class which takes in collision info then based on the objects the info contains calls a method with the suitable response

### CollisionInfo
Added methods to grab specific objects from the CollisionInfo based on type along with adding a method to see what objectTypes were invovled in the collision

### CollisionObject
Added a new enum so that each CollisionObject keeps track of what type of object it is.

Also allowed for the Direction to be properly set

### CollisionRect
Added rudimentary documentation for it's public method

### Game1
Reorganized the local variables of Game1 so there's some commentation

Added AllCollisionHandler to game1 which is intialized in Initialize

Added _movingCollisionObjects and _nonMovingCollisionObjects which are lists to keep track of collisionObjects. Currently only movingCollisionObject has the Player stored.

Added the HandleCollisions method which is called in Update which reads through both collisionLists to produce Collision Info which is then passed to AllCollisionHandler

### Player
Removed Position and Collider properties because they are just inherited from CollisionObject

### Util
Modified Util to account for changing of CollisionDirection Enum scope.