#41 #25 Created new wall class
### AllCollisionHandler
Removed debug statement

### CollisionObject
Objects were incorrectly determining collisionDirection, fixed that

### Game1
Removed singleton for now, should be a group decision

### Tile
Made it no longer a collisionObject to align with the vision

### Wall
A simple new class which serves as an immovable collision object which has the type of wall

### Player
Moved starting position to (200,200)

### PlayerCollisionHandler
Changed pushPixels to equal 2 so that walls actually stopped the player. PushPixels likely needs to become a variable which is always equal to the player's move speed

### Additional Notes
Collision Object currently has bad vertical collisions will work on that soon. If there's any odd behaviour vertically it's prolly cuz of that.