#15 Made some basic battle mechanics:
#42 Player and Enemies have a mechanical representation in their battle states

#43 Player and Enemies both have movesets to use during their respective turns

###EnemySet
To help keep track of the current enemies being fought in a battle, and also with who is currently alive.

###IEnemyBattle
Basic interface that each enemy shares to make implementation simpler

###Armos/Lynel?MoblinBattle
Battle representations of their respective character with each having unique and similar moves

###PlayerMoveset
Allows the player class to still be used, and gives the player a set moveset of four moves in battle. These being Attack, Defend, UseItem, and RunAway. Also keeps track as to whether the player's turn is active.

###PlayerStateMachine
Has a new currentDamageIntake that is adjusted when the player chooses to defend from enemy attacks.

#44 Made a command that can be used by the RPG UI

###GetPlayerStatsCommand
Currently returns the player's current health.
Needs to be updated to fit a sprite representation.