#71 Changed room 0-0 to have a different vibe
### Commands
Reorganized them into folders because they were getting a bit lengthy.

### Level0-0
Changed it to serve more like an entrance room, could move this to another level if we want a a dif vibe.

### test_tiles.png
Changed the tile map corners to be white

### Game folder
Renamed game folder to game states folder so it had a clearer name

### Entrance
Made it actually have the type of entrance

### IPlayer
Was removed because it's pointless

### PlayerCollisionHandler
Removed due to being useless

### Player
Changed starting Pos to (300,300) along with changing the magic number 3 in it's collider scaling to Util.GlobalScale

### AnimatableSprite
Made Width and Height get affected by global scale so that it accurately reported size

### Util
Changed GlobalScale to an int because most of the time it's being used with Sprites which are measured in ints.