#71 Added another room and reorganized rooms
### CharacterEnemyBase
Increased speed to match new scale

### EnemyWallCollideCommand
Made the code scale off of enemySpeed so that it dynamicall changes

### PlayerEntranceCollideCommand
Made it so that the newRoom adds the Player to it's movingCollisionList. Might break still, can't test due to room swapping being weird.

### PlayerWallCollide
Made Player use a property rather than a getter command

### Levels
Changed level values to be how it should. So now going down actually takes you to the room expected.

### Player
Gave player a MovementSpeed property with a public get