#91 Fixed Player Collision with Room Transitions
### Entrance
Made it so that when entrance changes room, it adds Player to that room's collisionList if it isn't there already.

Made ChangeRoom a static method so that it can be used by ChangeCurrentRoomCommand as well

### OverworldState
Put collisionDetection of moving and nonmoving into a try catch to shush away the index error which happens sometimes. That's uh risky I think

### PlayerEntranceCollideCommand
Removed global variables since they weren't doin nothin, and change it to make use of Entrance's static function call

### ChangeCurrentRoomCommand
Made ChangeCurrentRoomCommand use Entrance's static func