#116 New Weapon Pickup
### Level 0-3
Updates debug room with more funny things which have happened

### Inventory
Added new inventory representations of beamSword, boomerang, and bow

### WeaponOverworld
The overworld representation of weapons. Once picked up they add themselves to the inventory, granting access to using them.

### Map Parser
Allowed for it to parse weapons

### PlayerOverworld
Limited weapon usage based on if it's in the inventory or not